Add files via upload
Profit Engine NFTs were a 6700 NFT airdrop consisting of 6 different types that was completed on BSC to 2300 different holders. The NFTs had 6 nominal amounts contained within them that determined the NFTs % of total rewards collected. A mint dApp was included as well where NFTs were purchased in USDT. 

The owner of the contract could distribute USDT rewards to all 2300 holders for less than 5 cents, and holders could claim their rewards for less than 3 cents. 

I have no affiliation with the management of the Profit Engine project and was contracted to build their NFTs.