Add files via upload
This contract allows for affiliate marketing of a single token. Affiliates can have tokens deposited into a coupon with a set bonus percentage, and then people can buy them through a front end dApp. 

The bonus tokens are sent to the buyer immediately and a portion of the ETH used to buy the tokens is sent to the affiliate.