Add files via upload
Shillitary Testnet. 

Allows for project managers to pay a small fee to have tweets submitted to a raid army "The Shillitary". Users require an NFT badge that is airdropped to them to prevent spamming of the points by fraudulent individuals. 

Users get points based on number of tweets raided. An honour system is employed, and if broken, their badge is revokable for misconduct. 

Points are multiplied by owning a Frog Army NFT. Rewards are deposited in ETH and DeFi Tokens. 

Additional Claim function is still required, this commit is mainly for the point system.