Add files via upload
ERC-721 Contract allowing for group buys to be pooled together. 
The shared positions are managed by a python script set with different rules and was integrated into a TG Bot. 
Profitability of managers is tracked and once success is proven they can upgrade to "Mastermind" status where they get a cut of the profit of each successful trade.