Add files via upload
The UnityBot V1 Contract proved to have difficulties with group investments in tokens that offered rewards, and many investors liked the group buys, but did not like that they did not get their tokens right away. 

This modified contract allows for the the funds to be distributed right away and no longer tracks profitability.